# Switch to strconv.Parse* functions instead of fmt.Sscanf