--- a/AUTHORS
+++ b/AUTHORS
@@ -1,8 +1,10 @@
-There are several persons contribute their code include small patches
+There are many persons contribute their code include small patches
 to the project. They listed below in an alphabetical order:
 
 - Alexander (Axel) I.Grafov <grafov@gmail.com>
 - Andrey Chernov <chernov@bradburylab.com>
+- Andrew Sinclair <ajsinclair@gmail.com>
+- Bradley Falzon <brad@teambrad.net>
 - Denys Smirnov <denis.smirnov.91@gmail.com>
 - Fabrizio (Misto) Milo <mistobaan@gmail.com>
 - Jamie Stackhouse <jamie.stackhouse@redspace.com>