- Switch to strconv.Parse\* functions instead of fmt.Sscanf

Before:

```
BenchmarkDecodeMasterPlaylist-8    30000         45795 ns/op       12197 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-8       100      17158205 ns/op    11000952 B/op     120104 allocs/op
BenchmarkEncodeMasterPlaylist-8  1000000          1062 ns/op          40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-8       300       4762959 ns/op       45426 B/op        405 allocs/op
```

After:

```
BenchmarkDecodeMasterPlaylist-8    30000         41889 ns/op       11944 B/op         73 allocs/op
BenchmarkDecodeMediaPlaylist-8       100      16865034 ns/op    10998960 B/op     120091 allocs/op
BenchmarkEncodeMasterPlaylist-8  1000000          1091 ns/op          40 B/op         11 allocs/op
BenchmarkEncodeMediaPlaylist-8       300       4831626 ns/op       45419 B/op        405 allocs/op
```

Benchstat Summary:

```
name                    old time/op    new time/op    delta
DecodeMasterPlaylist-8    45.8µs ± 0%    41.9µs ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     17.2ms ± 0%    16.9ms ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8    1.06µs ± 0%    1.09µs ± 0%   ~             (p=1.000 n=1+1)
EncodeMediaPlaylist-8     4.76ms ± 0%    4.83ms ± 0%   ~             (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
DecodeMasterPlaylist-8    12.2kB ± 0%    11.9kB ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     11.0MB ± 0%    11.0MB ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8     40.0B ± 0%     40.0B ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8     45.4kB ± 0%    45.4kB ± 0%   ~             (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
DecodeMasterPlaylist-8      77.0 ± 0%      73.0 ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8       120k ± 0%      120k ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8      11.0 ± 0%      11.0 ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8        405 ± 0%       405 ± 0%   ~     (all samples are equal)
```